fix : changement des nom des composants
