+++ frontend/src/components/AnimatedRoutes/AnimatedRoutes.jsx
@@ -15,7 +15,7 @@ import About from "../../pages/About/About.jsx";
 import Dashboard from '../../pages/Dashboardservice/Dashboardservice';
 import Service from "../../pages/Service/Service.jsx"
 import Logout from '../../components/Logout/Logout';
-import CardList from '../../components/Listedescategorie/CardList';
+import CardList from '../ListedesCategorie/CardCategorie';
 import CardDetail from '../../components/DetailsDeService/CardDetails';
 import Contacter from '../../components/ContacterMessagerie/Contacter';
