--- frontend/src/components/ListedesCategorie/card.jsx
@@ -1,11 +1,0 @@
-import React from 'react';
-import CardList from './CardList'; // Ensure correct path
-
-const Card = () => {
-
-  return (
-    <CardList/>
-  );
-};
-
-export default Card;
